+++ p2panda-group/src/lib.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 pub mod crypto;
-pub mod key_bundle;
+pub mod keybundle;
 pub mod traits;
-pub mod two_party;
+pub mod twoparty;
