+++ p2panda-group/src/traits/keybundle.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
 use crate::crypto::x25519::PublicKey;
-use crate::key_bundle::{KeyBundleError, OneTimeKeyId};
+use crate::keybundle::{KeyBundleError, OneTimeKeyId};
 
 pub trait KeyBundle {
     fn identity_key(&self) -> &PublicKey;
